Cleaned and updated the insignia  module
