+++ addons/insignia/CfgUnitInsignia.hpp
@@ -1,12 +1,3 @@
-
-#define MACRO_UNITINSIGNIA(var1) \
-class TRIPLES(cav,insignia,var1) { \
-    displayName = CSTRING(var1); \
-    author = ECSTRING(Main,ModTeam); \
-    texture = QPATHTOF(data\var1.paa); \
-    textureVehicle = ""; \
-}
-
 class CfgUnitInsignia {
     // Alpha first platoon
     MACRO_UNITINSIGNIA(alpha_1);
