+++ addons/insignia/script_component.hpp
@@ -3,3 +3,4 @@
 
 #include "\z\cav\addons\main\script_mod.hpp"
 #include "\z\cav\addons\main\script_macros.hpp"
+#include "script_macros.hpp"
